[rename🚚] community도메인에서 image도메인 분리(#21)
