+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/image/repository/ImageRepository.java
@@ -1,6 +1,6 @@
-package com.culturespot.culturespotdomain.core.community.repository;
+package com.culturespot.culturespotdomain.core.image.repository;
 
-import com.culturespot.culturespotdomain.core.community.entity.Image;
+import com.culturespot.culturespotdomain.core.image.entity.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ImageRepository extends JpaRepository<Image, Long> {
