+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/image/service/ImageService.java
@@ -1,6 +1,6 @@
-package com.culturespot.culturespotdomain.core.community.service;
+package com.culturespot.culturespotdomain.core.image.service;
 
-import com.culturespot.culturespotdomain.core.community.entity.Image;
+import com.culturespot.culturespotdomain.core.image.entity.Image;
 import com.culturespot.culturespotdomain.core.post.entity.Post;
 import org.springframework.web.multipart.MultipartFile;
